Implement ELF loading functionality and associated tests
- Added `elf_loader` module to handle ELF file parsing and loading into task memory.
- Created `ElfHeader` and `ProgramHeader` structs for representing ELF headers and program headers.
- Implemented functions to read ELF headers and program headers with endian-aware reading.
- Developed `load_elf_into_task` function to load ELF files into a task's memory space.
- Introduced a mock ELF file for testing purposes.
- Added tests for loading an embedded ELF file into a task, verifying entry point and memory mapping.
- Updated `Cargo.toml` to include a new binary target for testing.
- Created a simple test binary that triggers an ECALL for testing purposes.
