+++ user/bin/Cargo.toml
@@ -7,5 +7,9 @@ edition = "2024"
 name = "init"
 path = "src/init.rs"
 
+[[bin]]
+name = "test"
+path = "src/test.rs"
+
 [dependencies]
 scarlet_std = { path = "../lib" }
